fix: stabilize v0.13.0 for develop (CodeQL, scan, complexity)
- Replace ReDoS-prone NON_PRODUCTION_CREDENTIAL_RE with linear parser (CodeQL)
- Extract ecosystem audit runners to audit-ecosystem.ts (file size)
- Fix complexity false positive in js-import-aliases brace parsing
- Resolve knip noise: internal stripJsonComments, single detectInstallChannel export
- Remove unused test import; fix eslint escapes in comments.ts
